README_JMC_DESC
This project was done in college and created mostly from scratch (the logos of different systems are not mine). It combines two things I love: coffee and gaming. This is to show that I know how to set up a website with two nav bars and a few working links with layered web pages. The concept is all mine and something I wouldn't mind actually seeing come to life at some point.